Moved FacetData and Ensemble data to sofast
