+++ opencsp/app/sofast/test/test_integration_multi_facet.py
@@ -11,8 +11,8 @@ from opencsp.app.sofast.lib.ProcessSofastFringe import ProcessSofastFringe as So
 from opencsp.app.sofast.lib.MeasurementSofastFringe import MeasurementSofastFringe as Measurement
 from opencsp.common.lib.camera.Camera import Camera
 from opencsp.app.sofast.lib.Display import Display
-from opencsp.common.lib.deflectometry.EnsembleData import EnsembleData
-from opencsp.common.lib.deflectometry.FacetData import FacetData
+from opencsp.app.sofast.lib.EnsembleData import EnsembleData
+from opencsp.app.sofast.lib.FacetData import FacetData
 from opencsp.common.lib.geometry.Vxyz import Vxyz
 from opencsp.common.lib.tool.hdf5_tools import load_hdf5_datasets
 from opencsp.common.lib.opencsp_path.opencsp_root_path import opencsp_code_dir
